restructured project with scenes in mind
