+++ index.js
@@ -8,16 +8,16 @@ import { name as appName } from "./app.json";
 import { createStore, applyMiddleware } from "redux";
 import { composeWithDevTools } from "redux-devtools-extension";
 import { Provider } from "react-redux";
-import rootReducer from "./src/reducers";
+import rootReducer from "./src/common/reducers";
 import createSagaMiddleware from "redux-saga";
 
 // Sagas
-import allSagas from "./src/sagas/sagas";
+import sagas from "./src/common/sagas";
 
 const composeEnhancers = composeWithDevTools({ realtime: true, port: 8081 });
 const sagaMiddleware = createSagaMiddleware();
 const store = createStore(rootReducer, composeEnhancers(applyMiddleware(sagaMiddleware)));
-sagaMiddleware.run(allSagas);
+sagaMiddleware.run(sagas);
 
 // App stuffs
 const ReduxApp = () => (
