+++ src/common/sagas/index.js
@@ -1,5 +1,5 @@
 import { fork } from "redux-saga/effects";
-import { authSagas } from "./auth";
+import { authSagas } from "../../scenes/auth/sagas/auth";
 
 export default function* root() {
   yield [fork(authSagas)];
