+++ src/scenes/AuthenticatedView.js
