+++ src/scenes/auth/signup/SignUpForm.js
@@ -1,41 +1,45 @@
-import * as React from "react";
-import {reduxForm, Field} from "redux-form";
-import {
-    ScrollView,
-    View,
-    Text,
-    TouchableOpacity,
-    StyleSheet
-} from "react-native";
+import React from "react";
+import CustomTextInput from "../../../common/components/CustomTextInput";
+import {ScrollView, StyleSheet, Text, TouchableOpacity, View} from "react-native";
+import {Field, reduxForm} from "redux-form";
+import {validateSignUp} from "./validators";
 
-import CustomTextInput from "../components/CustomTextInput";
-
-const SignInForm = props => {
+const SignUpForm = (props) => {
     return (
         <View>
-            <ScrollView keyboardShouldPersistTaps={"handled"}>
+            <ScrollView>
                 <Field
                     name={"email"}
                     component={CustomTextInput}
-                    placeholder={"Email"}
+                    placeholder={"your@email.here"}
                     style={styles.textField}
                 />
                 <Field
-                    name={"password"}
+                    name={"password1"}
                     component={CustomTextInput}
-                    placeholder={"Password"}
+                    placeholder={"Enter a password"}
+                    style={styles.textField}
+                    secureTextEntry
+                />
+                <Field
+                    name={"password2"}
+                    component={CustomTextInput}
+                    placeholder={"Re-enter your password"}
                     style={styles.textField}
                     secureTextEntry
                 />
                 <TouchableOpacity onPress={props.handleSubmit}>
-                    <Text style={{color: "#FFFFFF"}}>Sign In</Text>
+                    <Text style={{color: "#FFFFFF"}}>Sign Up</Text>
                 </TouchableOpacity>
             </ScrollView>
         </View>
     );
 };
 
-export default reduxForm({form: "signInForm"})(SignInForm);
+export default reduxForm({
+    form: "signUpForm",
+    validate: validateSignUp, // Sync validation
+})(SignUpForm);
 
 const styles = StyleSheet.create({
     container: {
